Working time in seconds from epoch

+++ include/EVT/dev/platform/f3xx/f302x8/RTC302x8.hpp
@@ -34,7 +34,7 @@ public:
      *
      * @return The time since epoch as determined by the RTC
      */
-    uint64_t getTime();
+    uint32_t getTime();
 
     /**
      * Set the time of the real time clock.
